chore: move authorization to factories and models
