+++ src/api/models/headers.api.model.ts
@@ -1,0 +1,3 @@
+export interface Headers {
+  [key: string]: string;
+}
