+++ tests/api/articles.integration.spec.ts
@@ -2,11 +2,9 @@ import {
   ArticlePayload,
   prepareArticlePayload,
 } from '@_src/api/factories/article-payload.api.factory';
-import {
-  Headers,
-  apiEndpoints,
-  getAuthorizationHeader,
-} from '@_src/api/utils/api.util';
+import { getAuthorizationHeader } from '@_src/api/factories/authorization-header.api.factory';
+import { Headers } from '@_src/api/models/headers.api.model';
+import { apiEndpoints } from '@_src/api/utils/api.util';
 import { expect, test } from '@_src/ui/fixtures/merge.fixture';
 import { APIResponse } from '@playwright/test';
