+++ tests/api/comments.integration.spec.ts
@@ -1,13 +1,11 @@
 import { prepareArticlePayload } from '@_src/api/factories/article-payload.api.factory';
+import { getAuthorizationHeader } from '@_src/api/factories/authorization-header.api.factory';
 import {
   CommentPayload,
   prepareCommentPayload,
 } from '@_src/api/factories/comment-payload.api.factory';
-import {
-  Headers,
-  apiEndpoints,
-  getAuthorizationHeader,
-} from '@_src/api/utils/api.util';
+import { Headers } from '@_src/api/models/headers.api.model';
+import { apiEndpoints } from '@_src/api/utils/api.util';
 import { expect, test } from '@_src/ui/fixtures/merge.fixture';
 import { APIResponse } from '@playwright/test';
